+++ quimb/tensor/belief_propagation/d2bp.py
@@ -847,6 +847,7 @@ class D2BP(BeliefPropagationCommon):
 
         return process_loop_series_expansion_weights(
             weights,
+            num_tensors=self.tn.num_tensors,
             mantissa=self.sign,
             exponent=self.exponent,
             multi_excitation_correct=multi_excitation_correct,
@@ -955,7 +956,9 @@ class D2BP(BeliefPropagationCommon):
             weights.pop(r0)
             # compute exponential suppresion factors
             corrections = process_loop_series_expansion_weights(
-                weights, return_all=True
+                weights,
+                num_tensors=self.tn.num_tensors,
+                return_all=True,
             )
             # add back in the BP contribution
             corrections[r0] = 1.0